DBA_Role_Privs granted

+++ db/migrate/20200625010000_create_allowed_db_tables.rb
@@ -38,13 +38,13 @@ class CreateAllowedDbTables < ActiveRecord::Migration[6.0]
         FROM   DBA_Tables t
         CROSS JOIN (SELECT Grantee FROM DBA_Sys_Privs WHERE Privilege = 'SELECT ANY TABLE') p
         "
-    else
+    when 'SQLITE' then
+      ActiveRecord::Base.connection.execute "DROP VIEW Allowed_DB_Tables" if view_exists? 'Allowed_DB_Tables'
       ActiveRecord::Base.connection.execute "CREATE VIEW Allowed_DB_Tables AS SELECT 'main' Owner, 'main' Grantee, Name Table_Name FROM SQLite_Master"
     end
   end
 
   def down
-    ActiveRecord::Base.connection.execute "DROP VIEW Allowed_DB_Tables"
   end
 end
